+++ src/pages/ErrorPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ErrorPage = () => {
+  return <div>잘못된 페이지입니다.</div>;
+};
+
+export default ErrorPage;
